requirements.txt
asgiref==3.5.2
Django==4.0.4
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-js-asset==2.0.0
django-resized==1.0.2
django-smartfields==1.1.3
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.9.5
pytz==2023.3
six==1.16.0
sqlparse==0.4.2
whitenoise==6.2.0
